# contracts/openapi.yaml
openapi: 3.1.1
...
info:
  title: 晶选片场 R0.6 Platform Foundation API
  version: 0.2.0-rc.1
  description: Account and party HTTP implementation candidate. IMPLEMENTED_ISOLATED_TESTS operations run on the
    separate MySQL account API. SMS real-provider verification and client integration are not complete; other operations
    remain NOT_IMPLEMENTED.
servers:
- url: http://127.0.0.1:3000
  description: Account API development entry; requires explicit MySQL configuration, applied migrations and secret.
    Not a production service.
security:
- bearerAuth: []
...
paths:
  /api/v1/auth/sms-challenges:
    post:
      operationId: createSmsChallenge
      summary: 请求登录短信
      description: No production dev-code fallback. Provider unavailable ->503, rate limit ->429. A retry with the
        same key must not resend SMS. New challenge creation invalidates only by explicitly documented server policy.
      tags:
      - session
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Success; only after real validation/committed work.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmsChallengeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmsChallengeRequest'
  /api/v1/auth/sessions:
    post:
      operationId: createSession
      summary: 验证码登录
      description: Bind challenge to normalized phone, expiry, one-time consumption and attempts. Invalid phone/challenge/code
        yields generic401 to resist enumeration. Commit challenge consumption + session + idempotency result atomically.
        Exact successful retry may return the original protected token result until session/result expiry; never
        mint a second session. No refresh or token storage policy implied.
      tags:
      - session
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Success; only after real validation/committed work.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionRequest'
  /api/v1/me:
    get:
      operationId: getCurrentAccount
      summary: 当前登录账户
      description: Bearer identity only; acting_party=null. Phone login is not real-name verification. A suspended
        account may read its own status but has only server-allowed actions.
      tags:
      - session
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Success; only after real validation/committed work.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/me/parties:
    get:
      operationId: listMyParties
      summary: 可访问主体与成员关系
      description: Only caller memberships. Invitations may be displayed but do not confer action authority. Select
        party client-side and send X-Acting-Party per request; do not mutate a global selected party in the bearer
        token.
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Success; only after real validation/committed work.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/parties/{party_id}:
    get:
      operationId: getParty
      summary: 主体详情
      description: Require active membership or explicit audited delegation. Route/header mismatch ->403; inaccessible
        private object ->404. Capabilities do not imply verified identity.
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ActingParty'
      - $ref: '#/components/parameters/party_id'
      responses:
        '200':
          description: Success; only after real validation/committed work.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
    patch:
      operationId: updatePartyDisplayName
      summary: 更新主体展示名称
      description: Only party.profile.update may change display_name. This is not legal-name/identity change or
        party transfer. Every successful new mutation increments object_version exactly once. Same idempotent request
        replays original result without a new increment. Do not use a read-time allowed_actions list instead of
        server authorization.
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/ActingParty'
      - $ref: '#/components/parameters/party_id'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Success; only after real validation/committed work.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyPatch'
...
  /api/v1/auth/sessions/current:
    delete:
      operationId: revokeCurrentSession
      summary: 退出当前登录
      tags:
      - session
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionRevokedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyRequest'
  /api/v1/organizations:
    post:
      operationId: createOrganization
      summary: 创建待审机构
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateRequest'
  /api/v1/me/invitations:
    get:
      operationId: listMyInvitations
      summary: 发给本人的邀请
      tags:
      - session
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/parties/{party_id}/capabilities:
    post:
      operationId: requestPartyCapability
      summary: 申请机构或个人能力
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - in: path
        name: party_id
        required: true
        schema: &id001
          $ref: '#/components/schemas/Id'
      - $ref: '#/components/parameters/ActingParty'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilityGrantResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapabilityRequest'
  /api/v1/parties/{party_id}/members:
    get:
      operationId: listPartyMembers
      summary: 负责人查看本机构成员
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - in: path
        name: party_id
        required: true
        schema: *id001
      - $ref: '#/components/parameters/ActingParty'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/v1/parties/{party_id}/members/{account_id}:
    delete:
      operationId: removePartyMember
      summary: 移除本机构普通成员
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - in: path
        name: party_id
        required: true
        schema: *id001
      - in: path
        name: account_id
        required: true
        schema: *id001
      - $ref: '#/components/parameters/ActingParty'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberRemovedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyRequest'
  /api/v1/parties/{party_id}/invitations:
    post:
      operationId: invitePartyMember
      summary: 邀请指定账号加入机构
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - in: path
        name: party_id
        required: true
        schema: *id001
      - $ref: '#/components/parameters/ActingParty'
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationCreatedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationCreateRequest'
  /api/v1/parties/{party_id}/invitations/{invitation_id}/responses:
    post:
      operationId: respondPartyInvitation
      summary: 本人接受或拒绝邀请
      tags:
      - party
      x-implementation-status: IMPLEMENTED_ISOLATED_TESTS
      parameters:
      - $ref: '#/components/parameters/RequestId'
      - in: path
        name: party_id
        required: true
        schema: *id001
      - in: path
        name: invitation_id
        required: true
        schema: *id001
      - $ref: '#/components/parameters/ActingParty'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Committed result; synthetic provider results are test-only.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationResponseResultResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvitationDecisionRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Server-validated account identity, issuer/audience/expiry/revocation. Account token alone does
        not authorize acting_party or platform administration.
  parameters:
    RequestId:
      in: header
      name: X-Request-Id
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9._:-]+$
      description: Optional correlation hint; server validates or replaces. Response header must equal meta.request_id.
        Do not use for idempotency.
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[A-Za-z0-9._:-]+$
      description: idempotency_key. Scope=(account or pre-auth subject, acting_party, method, canonical path, key).
        Persist normalized request fingerprint and committed result atomically. Same payload replays original result;
        changed payload/in-flight conflict gives 409. Re-authorize before replay; never include secrets in fingerprint
        logs.
...
  headers:
    RequestId:
      required: true
      description: Matches meta.request_id
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    NoStore:
      required: true
      description: Authenticated, token and private metadata responses are not cacheable.
      schema:
        type: string
        const: no-store
  responses:
    BadRequest:
      description: VALIDATION_ERROR / INVALID_CURSOR / PARTY_CONTEXT_REQUIRED
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: IDEMPOTENCY_KEY_REUSED / IDEMPOTENCY_IN_PROGRESS
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SmsChallengeRequest:
      type: object
      additionalProperties: false
      required:
      - phone
      - purpose
      properties:
        phone:
          type: string
          pattern: ^1[3-9][0-9]{9}$
        purpose:
          type: string
          enum:
          - LOGIN
      description: First slice retains mainland mobile scope of legacy route; never accepts a client-supplied verification
        result.
...
    SmsChallengeResponse:
      type: object
      additionalProperties: false
      required:
      - meta
      - data
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          $ref: '#/components/schemas/SmsChallenge'
...